+++ test/TaskSetTest.ts
@@ -35,6 +35,16 @@ suite("TaskSet", function TaskSetTest() {
     }
 
 
+    function startTasks(taskSet: TaskSet<string, string>, namePrefix: string, resMsgPrefix: string, count: number): Task<string, string>[] {
+        var tasks: Task<string, string>[] = [];
+        for (var i = 0; i < count; i++) {
+            var t = taskSet.startTask(namePrefix + i, Math.random() < 0.5 ? createTaskRes1(resMsgPrefix + i) : createTaskRes2(resMsgPrefix + i, Math.round(Math.random() * 10)));
+            tasks.push(t);
+        }
+        return tasks;
+    }
+
+
     test("task-set-success", function taskSetSuccessTest(done) {
         // test success
         var taskSet = new TaskSet<string, string>();
@@ -70,6 +80,34 @@ suite("TaskSet", function TaskSetTest() {
             done();
         }, function (err) {
             asr.equal(true, err == "error-1" || err == "error-2");
+            done();
+        });
+    });
+
+
+    test("task-drop-completed", function taskSetDropCompleted(done) {
+        var taskSet = new TaskSet<string, string>();
+        // test 4 tasks, limit 3, drop 25%
+        taskSet.maxCompletedTasks = 3;
+        taskSet.dropCompletedTasksPercentage = 0.25;
+        startTasks(taskSet, "task-res-", "success-", 4);
+
+        Q.all(taskSet.getPromises()).then(function (res) {
+            asr.equal(taskSet.getCompletedTasks().length, 2);
+            taskSet.clearCompletedTasks();
+        }).then(function () {
+            // test 6 tasks, limit 5, drop 60%
+            taskSet.maxCompletedTasks = 5;
+            taskSet.dropCompletedTasksPercentage = 0.6;
+            startTasks(taskSet, "task-res-", "success-", 6);
+
+            return Q.all(taskSet.getPromises());
+        }).done(function (res) {
+            asr.equal(taskSet.getCompletedTasks().length, 2);
+            taskSet.clearCompletedTasks();
+            done();
+        }, function (err) {
+            asr.equal(false, true, "unexpected error");
             done();
         });
     });